training/requirements.txt
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0
peft>=0.6.0
trl>=0.7.0
wandb>=0.16.0
psycopg2-binary>=2.9.7
redis>=5.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.14.0
deepspeed>=0.11.0
bitsandbytes>=0.41.0
sentencepiece>=0.1.99
protobuf>=4.24.0
